+++ pysrc/juliacall/__init__.py
@@ -1,7 +1,7 @@
 # This module gets modified by PythonCall when it is loaded, e.g. to include Core, Base
 # and Main modules.
 
-__version__ = '0.9.33'
+__version__ = '0.9.34'
 
 _newmodule = None
 
@@ -170,39 +170,47 @@ def init():
     CONFIG['opt_handle_signals'] = choice('handle_signals', ['yes', 'no'])[0]
     CONFIG['opt_startup_file'] = choice('startup_file', ['yes', 'no'])[0]
     CONFIG['opt_heap_size_hint'] = option('heap_size_hint')[0]
-    CONFIG['project'] = path_option('project', check_exists=True)[0]
-    CONFIG['exepath'] = executable_option('exe')[0]
+    CONFIG['project'] = project = path_option('project', check_exists=True)[0]
+    CONFIG['libpath'] = libpath = path_option('lib', check_exists=True)[0]
+    CONFIG['exepath'] = exepath = executable_option('exe')[0]
 
     # Stop if we already initialised
     if CONFIG['inited']:
         return
 
-    have_exepath = CONFIG['exepath'] is not None
-    have_project = CONFIG['project'] is not None
-    if have_exepath and have_project:
+    if (exepath is None) and (bindir is not None):
+        # if bindir is set then set exepath={bindir}/julia
+        CONFIG['exepath'] = exepath = os.path.join(bindir, 'julia.exe' if os.name == 'nt' else 'julia')
+    if (exepath is not None) and (project is not None):
         pass
-    elif (not have_exepath) and (not have_project):
+    elif (exepath is None) and (project is None):
         # we don't import this at the top level because it is not required when
         # juliacall is loaded by PythonCall and won't be available, or if both
         # `exepath` and `project` are set by the user.
         import juliapkg
 
         # Find the Julia executable and project
-        CONFIG['exepath'] = juliapkg.executable()
-        CONFIG['project'] = juliapkg.project()
+        CONFIG['exepath'] = exepath = juliapkg.executable()
+        CONFIG['project'] = project = juliapkg.project()
+        if libpath is None:
+            CONFIG['libpath'] = libpath = juliapkg.libjulia()
     else:
         raise Exception("Both PYTHON_JULIACALL_PROJECT and PYTHON_JULIACALL_EXE must be set together, not only one of them.")
+    if (libpath is not None) and (exepath is None):
+        raise Exception("PYTHON_JULIACALL_EXE is required if PYTHON_JULIACALL_LIB is set.")
 
-    exepath = CONFIG['exepath']
-    project = CONFIG['project']
-
-    # Find the Julia library
-    cmd = [exepath, '--project='+project, '--startup-file=no', '-O0', '--compile=min',
-           '-e', 'import Libdl; print(abspath(Libdl.dlpath("libjulia")), "\\0", Sys.BINDIR)']
-    libpath, default_bindir = subprocess.run(cmd, check=True, capture_output=True, encoding='utf8').stdout.split('\0')
+    # Find the Julia library, if not specified.
+    if libpath is None:
+        cmd = [exepath, '--project='+project, '--startup-file=no', '-O0', '--compile=min',
+               '-e', 'import Libdl; print(abspath(Libdl.dlpath("libjulia")), "\\0", Sys.BINDIR)']
+        libpath, found_bindir = subprocess.run(cmd, check=True, capture_output=True, encoding='utf8').stdout.split('\0')
+        CONFIG['libpath'] = libpath
+        if bindir is None:
+            CONFIG['bindir'] = bindir = found_bindir
+    if bindir is None:
+        bindir = os.path.dirname(exepath)
     assert os.path.exists(libpath)
-    assert os.path.exists(default_bindir)
-    CONFIG['libpath'] = libpath
+    assert os.path.exists(bindir)
 
     # Add the Julia library directory to the PATH on Windows so Julia's system libraries can
     # be found. They are normally found because they are in the same directory as julia.exe,
@@ -225,6 +233,11 @@ def init():
     jl_parse_opts(c.pointer(argc), c.pointer(argv))
     assert argc.value == 0
 
+    # override some environment variables
+    # we do this here because PythonCall is initialised during jl_init if it is in a sysimg
+    os.environ['JULIA_PYTHONCALL_EXECUTABLE'] = sys.executable or ''
+    os.environ['__JULIA_PYTHONCALL_EMBEDDED_LIBPTR__'] = hex(c.pythonapi._handle)
+
     # initialise julia
     try:
         jl_init = lib.jl_init_with_image__threading
@@ -233,7 +246,7 @@ def init():
     jl_init.argtypes = [c.c_char_p, c.c_char_p]
     jl_init.restype = None
     jl_init(
-        (default_bindir if bindir is None else bindir).encode('utf8'),
+        None if bindir is None else bindir.encode('utf8'),
         None if sysimg is None else sysimg.encode('utf8'),
     )
 
@@ -254,9 +267,7 @@ def init():
         return 'raw"' + x + '"'
     script = '''
     try
-        Base.require(Main, :CompilerSupportLibraries_jll)
-        global __PythonCall_libptr = Ptr{{Cvoid}}(UInt({}))
-        ENV["JULIA_PYTHONCALL_EXE"] = {}
+        import CompilerSupportLibraries_jll as _
         using PythonCall
     catch err
         print(stderr, "ERROR: ")
@@ -264,13 +275,14 @@ def init():
         flush(stderr)
         rethrow()
     end
-    '''.format(
-        hex(c.pythonapi._handle),
-        jlstr(sys.executable or ''),
-    )
+    '''
     res = jl_eval(script.encode('utf8'))
     if res is None:
         raise Exception('PythonCall.jl did not start properly')
+    
+    # unset this env var so any other julia processes started do not think they are
+    # embedded in python
+    os.environ.pop('__JULIA_PYTHONCALL_EMBEDDED_LIBPTR__', None)
 
     CONFIG['inited'] = True