+++ CHANGELOG.md
@@ -32,6 +32,15 @@
   * `pyjltype(x)` removed.
   * New functions: `pyjlarray`, `pyjldict`, `pyjlset`.
 
+## Unreleased
+* Add option `lib` to JuliaCall. Setting this will skip the discovery subprocess.
+* Add support for using a system image in `juliacall` that has `PythonCall` baked in.
+* Add method `pynext(x, d)` to return a default value `d` if there are no more elements.
+* Bug fixes.
+
+## 0.9.34 (2026-05-18)
+* Bug fixes.
+
 ## 0.9.33 (2026-05-18)
 * Add configuration via Preferences in addition to environment variables (e.g. `exe`
   rather than `JULIA_PYTHONCALL_EXE`.)